chore(users): move EmploymentType from Contracts to Enums namespace
Addresses Marcel's nit on PR #3653 — Contracts/ conventionally
implies interfaces, but EmploymentType is a value enum. Discoverability
is better under Users\Enums\ where PHP enums live throughout the
codebase.

Pure rename — no behavior change. Namespace updates in:
- app/Domain/Users/Enums/EmploymentType.php (declaration)
- app/Domain/Users/Repositories/Users.php (use import)
- app/Domain/Users/Templates/editUser.blade.php (FQN in @foreach)
- tests/Unit/app/Domain/Users/Enums/EmploymentTypeTest.php (namespace + use)
- tests/Unit/app/Domain/Users/Repositories/UsersRepositoryTest.php (use)

## Test run
- EmploymentTypeTest: 5 tests / 28 assertions — all pass.
- UsersRepositoryTest: 11 tests / 22 assertions — all pass.
